+++ index.html
@@ -6,8 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <scripts src ="/main.js"></scripts>
-  <script  src="js/functions.js"></script>
+  <script src ="js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
